Implement StatementInterface on ClassDependency
